refactor: consolidate self-host hero into primary hero component via variant prop
